requirements.txt
Flask>=2.3.2
flask-cors>=3.0.10
Flask-SQLAlchemy>=2.5.1
psycopg2-binary>=2.8.6
# flask_bcrypt